requirements.txt
# API y Cliente
anthropic==0.39.0
python-dotenv==1.0.0

# Framework Web
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0

# Procesamiento de Datos
pandas==2.2.0
numpy==1.26.0

# Machine Learning / RAG
sentence-transformers==3.3.0
torch==2.5.0

# OCR
pytesseract==0.3.13
Pillow==11.0.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
hypothesis==6.122.0
httpx==0.28.0
